Upgrade dependencies and restructure file tree
Additions
- Created run.bat file for Windows
- Moved package fileio under mayonez for parity with physics branch
