+++ build.gradle
@@ -1,9 +1,9 @@
 plugins {
     id 'java'
     id 'application'
-    id 'org.jetbrains.kotlin.jvm' version '1.5.31'
-    id 'org.jetbrains.dokka' version '1.5.30' // For Kotlin Documentation
-    id 'com.github.johnrengelman.shadow' version '7.0.0' // For making fat jars
+    id 'org.jetbrains.kotlin.jvm' version '1.6.0'
+    id 'org.jetbrains.dokka' version '1.6.0' // For Kotlin Documentation
+    id 'com.github.johnrengelman.shadow' version '7.1.0' // For making fat jars
 }
 
 group 'slavsquatsuperstar'
@@ -12,7 +12,7 @@ version '0.7.1-SNAPSHOT'
 repositories.mavenCentral()
 
 // LWJGL Settings
-project.ext.lwjglVersion = '3.2.3'
+project.ext.lwjglVersion = '3.3.0'
 project.ext.lwjglNatives = 'natives-macos'
 
 dependencies {
@@ -36,15 +36,15 @@ dependencies {
     implementation "org.joml:joml:1.10.2"
 
     // Code Dependencies
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.5.31'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.6.0'
     implementation 'org.apache.commons:commons-lang3:3.12.0'
     implementation 'commons-io:commons-io:2.11.0'
-    implementation 'org.json:json:20210307'
-    implementation 'org.reflections:reflections:0.9.12'
+    implementation 'org.json:json:20211205'
+    implementation 'org.reflections:reflections:0.10.2'
 
     // Test Libraries
-    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
+    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
 }
 
 String javaVersion = '11'
