+++ src/main/kotlin/slavsquatsuperstar/mayonez/fileio/Assets.kt
@@ -1,4 +1,4 @@
-package slavsquatsuperstar.fileio
+package slavsquatsuperstar.mayonez.fileio
 
 import org.reflections.Reflections
 import org.reflections.scanners.ResourcesScanner
